+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Move role cache data from SecureStore into json CacheStore #26
  * `exec` command will abort if a conflicting AWS Env var is set #27
  * Add `time` command to report how much time before the current STS token expires #28
+ * Add support for printing Arn in `list` #33
 
 ## [v1.0.1] - 2021-07-18